+++ backend/src/main/java/com/prograngers/backend/config/WebConfig.java
@@ -1,4 +1,4 @@
-package com.prograngers.backend;
+package com.prograngers.backend.config;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
