+++ backend/src/main/java/com/prograngers/backend/controller/auth/Login.java
@@ -1,4 +1,4 @@
-package com.prograngers.backend;
+package com.prograngers.backend.controller.auth;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
