+++ backend/src/main/java/com/prograngers/backend/exception/ErrorCode.java
@@ -15,6 +15,7 @@ public enum ErrorCode {
     FAIL_TO_ENCODED,
     FAIL_TO_DECODED,
     NOT_FOUND_REFRESH_TOKEN,
-    NOT_EXIST_REFRESH_TOKEN;
+    NOT_EXIST_REFRESH_TOKEN,
+    INCORRECT_PASSWORD;
 
 }
