--- backend/src/main/java/com/prograngers/backend/exception/InvalidException.java
@@ -1,8 +1,0 @@
-package com.prograngers.backend.exception;
-
-public class InvalidException extends CustomException {
-
-    public InvalidException(ErrorCode errorCode, String message) {
-        super(errorCode, message);
-    }
-}
