--- backend/src/main/java/com/prograngers/backend/exception/UnAuthorizationException.java
@@ -1,7 +1,0 @@
-package com.prograngers.backend.exception;
-
-public class UnAuthorizationException extends CustomException{
-    public UnAuthorizationException(ErrorCode errorCode, String message) {
-        super(errorCode, message);
-    }
-}
