+++ backend/src/main/java/com/prograngers/backend/exception/unauthorization/ExpiredTokenException.java
@@ -1,8 +1,8 @@
-package com.prograngers.backend.exception;
+package com.prograngers.backend.exception.unauthorization;
 
 import static com.prograngers.backend.exception.ErrorCode.EXPIRED_TOKEN;
 
-public class ExpiredTokenException extends UnAuthorizationException{
+public class ExpiredTokenException extends UnAuthorizationException {
     public ExpiredTokenException() {
         super(EXPIRED_TOKEN, "토큰의 유효기간이 만료되었습니다.");
     }
