+++ backend/src/main/java/com/prograngers/backend/exception/unauthorization/FailToEncodeException.java
@@ -1,6 +1,4 @@
-package com.prograngers.backend.dto;
-
-import com.prograngers.backend.exception.UnAuthorizationException;
+package com.prograngers.backend.exception.unauthorization;
 
 import static com.prograngers.backend.exception.ErrorCode.FAIL_TO_ENCODED;
