+++ backend/src/main/java/com/prograngers/backend/exception/unauthorization/RefreshTokenNotFoundException.java
@@ -1,4 +1,6 @@
-package com.prograngers.backend.exception;
+package com.prograngers.backend.exception.unauthorization;
+
+import com.prograngers.backend.exception.ErrorCode;
 
 public class RefreshTokenNotFoundException extends UnAuthorizationException {
     public RefreshTokenNotFoundException() {
