+++ backend/src/main/java/com/prograngers/backend/repository/MemberRepository.java
@@ -2,9 +2,11 @@ package com.prograngers.backend.repository;
 
 import com.prograngers.backend.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
+@Repository
 public interface MemberRepository extends JpaRepository<Member, Long> {
     Optional<Member> findByEmail(String email);
 }
