+++ backend/src/main/java/com/prograngers/backend/repository/RefreshTokenRepository.java
@@ -1,9 +1,11 @@
 package com.prograngers.backend.repository;
 
-import com.prograngers.backend.RefreshToken;
+import com.prograngers.backend.service.auth.RefreshToken;
 import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
+@Repository
 public interface RefreshTokenRepository extends CrudRepository<RefreshToken, Long> {
     Optional<RefreshToken> findById(Long memberId);
