+++ backend/src/main/java/com/prograngers/backend/service/auth/RefreshToken.java
@@ -1,4 +1,4 @@
-package com.prograngers.backend;
+package com.prograngers.backend.service.auth;
 
 import lombok.AccessLevel;
 import lombok.Builder;
